+++ src/actions/index.js
@@ -1,7 +1,7 @@
-export * from './GetMedia';
-export * from './PlayerFooter';
-export * from './PlaybackActions';
-export * from './PlaylistActions';
-export * from './FileSystem';
-export * from './Settings';
-export * from './GetLyrics';
+export * from './getMedia';
+export * from './playerFooter';
+export * from './playback';
+export * from './playlist';
+export * from './fileSystem';
+export * from './settings';
+export * from './getLyrics';
