+++ src/reducers/settings.js
