+++ src/utils/ErrorReporter.js
@@ -3,7 +3,7 @@ import { Alert, Linking } from 'react-native';
 export default function (e) {
 	Alert.alert(
 		'Oops! an error ocurred',
-		'Send crash logs to developers?',
+		'Send error log to developers?',
 		[{ text: 'Send', onPress: () => mailError(e) }],
 		{ cancelable: false }
 	);
